分片上传

+++ src/main/java/com/example/demo/upload/controller/FileController.java
@@ -4,7 +4,6 @@ import com.example.demo.upload.entity.FileDTO;
 import com.example.demo.upload.service.FileService;
 import com.example.demo.upload.utils.Result;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.io.FilenameUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
@@ -12,14 +11,16 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.*;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.IOException;
 import java.util.List;
-import java.util.UUID;
 
 /**
  * 分片上传 参考：
  * https://www.jb51.net/article/190808.htm
- *
+ * <p>
  * 访问路径：
  * http://localhost:8000/file/show
  */
@@ -68,11 +69,6 @@ public class FileController {
                          String key
     ) throws IOException, InterruptedException {
         log.info("上传文件开始");
-        //文件的名称
-        //String name = UUID.randomUUID().toString().replaceAll("-", "");
-        // 获取文件的扩展名
-        //String ext = FilenameUtils.getExtension(file.getOriginalFilename());
-
         //设置图片新的名字
         String fileName = new StringBuffer().append(key).append(".").append(suffix).toString(); // course\6sfSqfOwzmik4A4icMYuUe.mp4
         //这个是分片的名字
@@ -89,8 +85,7 @@ public class FileController {
         FileDTO file1 = new FileDTO();
         file1.setPath(basePath + localfileName);
         file1.setSuffix(suffix);
-        file1.setName(originName);
-        //file1.setSuffix(ext);
+        file1.setOriginName(originName);
         file1.setSize(size);
         file1.setShardIndex(shardIndex);
         file1.setShardSize(shardSize);
@@ -102,7 +97,7 @@ public class FileController {
         //判断当前是不是最后一个分页 如果不是就继续等待其他分页  合并分页
         if (shardIndex.equals(shardTotal)) {
             file1.setPath(basePath + fileName);
-            this.merge(file1);
+            this.mergeShards(file1);
         }
         return "上传成功";
     }
@@ -123,50 +118,52 @@ public class FileController {
      * @author fengxinglie
      * 合并分页
      */
-    private void merge(FileDTO fileDTO) throws FileNotFoundException, InterruptedException {
-        //合并分片开始
-        log.info("分片合并开始");
-        String path = fileDTO.getPath(); //获取到的路径 没有.1 .2 这样的东西
-        //截取视频所在的路径
-        path = path.replace(basePath, "");
-        Long shardTotal = fileDTO.getShardTotal();
-        File newFile = new File(basePath + fileDTO.getName());
-        FileOutputStream outputStream = new FileOutputStream(newFile, true); // 文件追加写入
-        FileInputStream fileInputStream = null; //分片文件
-        byte[] byt = new byte[10 * 1024 * 1024];
-        int len;
-        try {
-            for (int i = 0; i < shardTotal; i++) {
-                // 读取第i个分片
-                fileInputStream = new FileInputStream(new File(basePath + path + "." + (i + 1))); //  course\6sfSqfOwzmik4A4icMYuUe.mp4.1
-                while ((len = fileInputStream.read(byt)) != -1) {
-                    outputStream.write(byt, 0, len);
-                }
-            }
-        } catch (IOException e) {
-            log.error("分片合并异常", e);
-        } finally {
-            try {
-                if (fileInputStream != null) {
-                    fileInputStream.close();
-                }
-                outputStream.close();
-                log.info("IO流关闭");
-            } catch (Exception e) {
-                log.error("IO流关闭", e);
-            }
-        }
-        log.info("分片结束了");
+    private void mergeShards(FileDTO fileDTO) throws InterruptedException {
+        boolean toFile = shardsToFile(fileDTO);
+
         //告诉java虚拟机去回收垃圾 至于什么时候回收  这个取决于 虚拟机的决定
         System.gc();
         //等待100毫秒 等待垃圾回收去 回收完垃圾
         Thread.sleep(100);
+
+        if(toFile){
+            deleteShards(fileDTO);
+        }
+    }
+
+    private boolean shardsToFile(FileDTO fileDTO) {
+        log.info("分片合并开始");
+        File newFile = new File(basePath + fileDTO.getOriginName());
+        byte[] buffer = new byte[10 * 1024 * 1024];
+        int len;
+        try (FileOutputStream fos = new FileOutputStream(newFile, true)) {
+            for (int i = 0; i < fileDTO.getShardTotal(); i++) {
+                try (FileInputStream fis = new FileInputStream(new File(fileDTO.getPath() + "." + (i + 1)))) {
+                    while ((len = fis.read(buffer)) != -1) {
+                        fos.write(buffer, 0, len);
+                    }
+                }
+            }
+        } catch (IOException e) {
+            log.error("分片合并异常, error = {}", e);
+            if (newFile.exists()) {
+                newFile.delete();
+            }
+            return false;
+        }
+        log.info("分片结束了");
+        return true;
+    }
+
+    private void deleteShards(FileDTO fileDTO) {
         log.info("删除分片开始");
-        for (int i = 0; i < shardTotal; i++) {
-            String filePath = basePath + path + "." + (i + 1);
+        for (int i = 0; i < fileDTO.getShardTotal(); i++) {
+            String filePath = fileDTO.getPath() + "." + (i + 1);
             File file = new File(filePath);
-            boolean result = file.delete();
-            log.info("删除{}，{}", filePath, result ? "成功" : "失败");
+            if (file.exists()) {
+                boolean result = file.delete();
+                log.info("删除{}，{}", filePath, result ? "成功" : "失败");
+            }
         }
         log.info("删除分片结束");
     }
